Agent: Move TCPRelay into network.relay package
