+++ monkey/infection_monkey/telemetry/messengers/exploit_intercepting_telemetry_messenger.py
@@ -1,6 +1,6 @@
 from functools import singledispatch
 
-from infection_monkey.tcp_relay import TCPRelay
+from infection_monkey.network.relay.tcp_relay import TCPRelay
 from infection_monkey.telemetry.exploit_telem import ExploitTelem
 from infection_monkey.telemetry.i_telem import ITelem
 from infection_monkey.telemetry.messengers.i_telemetry_messenger import ITelemetryMessenger
